+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Feature:
   - Add a new workflow to push docker images for arm64 and amd64
 
+- Fixes:
+  - Fix sw8 loss when use aiohttp.(#299,issue#10669)
+
 ### 1.0.0
 
 - **Important Note and Breaking Changes:**
@@ -53,7 +56,6 @@
   - Add namespace suffix to service name (#275)
   - Add periodical instance property report to prevent data loss (#279)
   - Fix sw_logging when `Logger.disabled` is true (#281)
-  - Fix sw8 loss when use aiohttp.(#299,issue#10669)
 
 - Docs:
   - New documentation on how to test locally (#222)