+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Feature:
   - Add a new workflow to push docker images for arm64 and amd64
 
+- Fixes:
+  - Fix sw8 loss when use aiohttp.(#299,issue#10669)
+
 ### 1.0.0
 
 - **Important Note and Breaking Changes:**