+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ### 1.1.0
 
+- Feature:
+  - Add a new workflow to push docker images for arm64 and amd64
+
+- Fixes:
+  - Fix sw8 loss when use aiohttp.(#299,issue#10669)
+  - Fix the bug with high cpu usage.(#300,issue#10672)
+
 ### 1.0.0
 
 - **Important Note and Breaking Changes:**