+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Change Logs
 
+### 1.1.0
+
 ### 1.0.1
 
 - Feature: